refactor type organization
