+++ libs/fujitsu-thermal-printer/src/mocks/memory_printer.ts
@@ -2,8 +2,11 @@ import { tmpName } from 'tmp-promise';
 import { writeFile } from 'fs/promises';
 import { rmSync } from 'fs';
 import { ok } from '@votingworks/basics';
-import { PrinterStatus } from '../status';
-import { FujitsuThermalPrinterInterface, PrintResult } from '../printer';
+import {
+  FujitsuThermalPrinterInterface,
+  PrintResult,
+  PrinterStatus,
+} from '../types';
 
 /**
  * A mock of the  interface. See for details.
